requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
yfinance>=0.2.36
schedule>=1.2.0
click>=8.1.0
